LPD-25784 Move to the proper package
