+++ modules/apps/portal-scheduler/portal-scheduler-quartz/src/main/java/com/liferay/portal/scheduler/quartz/internal/upgrade/release/QuartzSchemaCreator.java
@@ -3,12 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.configuration.persistence.internal.upgrade.schema;
+package com.liferay.portal.scheduler.quartz.internal.upgrade.release;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.db.DBResourceUtil;
 import com.liferay.portal.kernel.dao.db.DB;
-import com.liferay.portal.kernel.dao.db.DBInspector;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.upgrade.UpgradeException;
@@ -25,35 +24,26 @@ import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Mariano Álvaro Sáiz
+ * @author Shuyang Zhou
  */
 @Component(service = SchemaCreator.class)
-public class ConfigurationSchemaCreator implements SchemaCreator {
+public class QuartzSchemaCreator implements SchemaCreator {
 
 	@Override
 	public void create() throws UpgradeException {
+		DB db = DBManagerUtil.getDB();
+
+		String indexesSQL = DBResourceUtil.getModuleIndexesSQL(_bundle);
+		String tablesSQL = DBResourceUtil.getModuleTablesSQL(_bundle);
+
 		try (Connection connection = DataAccess.getConnection()) {
-			DBInspector dbInspector = new DBInspector(connection);
+			db.runSQLTemplate(connection, tablesSQL, true);
 
-			if (dbInspector.hasTable("Configuration_")) {
-				return;
-			}
-
-			DB db = DBManagerUtil.getDB();
-
-			String tablesSQL = DBResourceUtil.getModuleTablesSQL(_bundle);
-
-			db.runSQLTemplate(connection, tablesSQL, false);
+			db.runSQLTemplate(connection, indexesSQL, true);
 		}
 		catch (Exception exception) {
 			throw new UpgradeException(exception);
 		}
-	}
-
-	public void create(Bundle bundle) throws UpgradeException {
-		_bundle = bundle;
-
-		create();
 	}
 
 	@Override
