+++ modules/apps/portal/portal-spring-extender-impl/src/main/java/com/liferay/portal/spring/extender/internal/upgrade/release/SchemaCreatorImpl.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.spring.extender.internal.release;
+package com.liferay.portal.spring.extender.internal.upgrade.release;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
