+++ modules/apps/static/portal-configuration/portal-configuration-persistence-impl/src/main/java/com/liferay/portal/configuration/persistence/internal/upgrade/release/ConfigurationSchemaCreator.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.scheduler.quartz.internal.upgrade.schema;
+package com.liferay.portal.configuration.persistence.internal.upgrade.release;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.db.DBResourceUtil;
 import com.liferay.portal.kernel.dao.db.DB;
+import com.liferay.portal.kernel.dao.db.DBInspector;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.upgrade.UpgradeException;
@@ -24,26 +25,35 @@ import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Shuyang Zhou
+ * @author Mariano Álvaro Sáiz
  */
 @Component(service = SchemaCreator.class)
-public class QuartzSchemaCreator implements SchemaCreator {
+public class ConfigurationSchemaCreator implements SchemaCreator {
 
 	@Override
 	public void create() throws UpgradeException {
-		DB db = DBManagerUtil.getDB();
-
-		String indexesSQL = DBResourceUtil.getModuleIndexesSQL(_bundle);
-		String tablesSQL = DBResourceUtil.getModuleTablesSQL(_bundle);
-
 		try (Connection connection = DataAccess.getConnection()) {
-			db.runSQLTemplate(connection, tablesSQL, true);
+			DBInspector dbInspector = new DBInspector(connection);
 
-			db.runSQLTemplate(connection, indexesSQL, true);
+			if (dbInspector.hasTable("Configuration_")) {
+				return;
+			}
+
+			DB db = DBManagerUtil.getDB();
+
+			String tablesSQL = DBResourceUtil.getModuleTablesSQL(_bundle);
+
+			db.runSQLTemplate(connection, tablesSQL, false);
 		}
 		catch (Exception exception) {
 			throw new UpgradeException(exception);
 		}
+	}
+
+	public void create(Bundle bundle) throws UpgradeException {
+		_bundle = bundle;
+
+		create();
 	}
 
 	@Override
